+++ mrcz/__init__.py
@@ -1,6 +1,6 @@
 
 from mrcz.ioMRC import (readMRC, writeMRC, asyncReadMRC, asyncWriteMRC, 
-    _setAsyncWorkers, _asyncExecutor )
+    _setAsyncWorkers, _asyncExecutor, setDefaultThreads)
 from mrcz.ioDM import readDM4
 from mrcz.__version__ import __version__
 from mrcz.test_mrcz import test